Add receipt repository and receipt service (#6)
- Add new dependencies
- Update receipt_handler.go with dependency injection
- Add mocks for receipt service and repository
- Update main.go to handle dependency injector
- Add receipt repository tests
- Add receipt service tests

+++ internal/dto/response.go
@@ -5,7 +5,7 @@ type CreateReceiptResponse struct {
 }
 
 type GetPointsResponse struct {
-	Points int64 `json:"points"`
+	Points int `json:"points"`
 }
 
 type ResponseErrorModel struct {
